Refactor reusable lsp4j utilities into a seperate library
