+++ language-server/langserver/pom.xml
@@ -2,8 +2,8 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
+        <groupId>edu.umn.cs.melt</groupId>
         <artifactId>silver-langserver-parent</artifactId>
-        <groupId>edu.umn.cs.melt.silver.langserver</groupId>
         <version>0.1.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
@@ -12,6 +12,11 @@
     <packaging>jar</packaging>
 
     <dependencies>
+        <dependency>
+            <groupId>edu.umn.cs.melt</groupId>
+            <artifactId>lsp4j-util</artifactId>
+            <version>0.1.0-SNAPSHOT</version>
+        </dependency>
         <dependency>
             <groupId>edu.umn.cs.melt</groupId>
             <artifactId>silver</artifactId>
