+++ language-server/langserver/src/main/java/edu/umn/cs/melt/silver/langserver/SilverLanguageService.java
@@ -65,6 +65,9 @@ import common.OriginContext;
 import common.SilverCopperParser;
 import common.StringCatter;
 import common.javainterop.ConsCellCollection;
+import edu.umn.cs.melt.lsp4jutil.CopperParserNodeFactory;
+import edu.umn.cs.melt.lsp4jutil.CopperSemanticTokenEncoder;
+import edu.umn.cs.melt.lsp4jutil.Util;
 import silver.compiler.composed.Default.Parser_silver_compiler_composed_Default_svParse;
 import silver.compiler.definition.core.NRoot;
 import silver.compiler.driver.PbuildRun;
