+++ language-server/launcher/pom.xml
@@ -3,7 +3,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <artifactId>silver-langserver-parent</artifactId>
-        <groupId>edu.umn.cs.melt.silver.langserver</groupId>
+        <groupId>edu.umn.cs.melt</groupId>
         <version>0.1.0-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
@@ -15,7 +15,7 @@
     <dependencies>
         <dependency>
             <artifactId>silver-langserver</artifactId>
-        	<groupId>edu.umn.cs.melt.silver.langserver</groupId>
+        	<groupId>edu.umn.cs.melt</groupId>
             <version>0.1.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
